# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PeriOp Voice API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /procedures:
    get:
      operationId: listProcedures
      tags: [procedures]
      summary: List procedures
      parameters:
        - in: query
          name: patientId
          schema: { type: integer }
        - in: query
          name: dateFrom
          schema: { type: string, format: date }
        - in: query
          name: dateTo
          schema: { type: string, format: date }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Procedure list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Procedure"
    post:
      operationId: createProcedure
      tags: [procedures]
      summary: Create procedure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcedureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
  /procedures/{id}:
    get:
      operationId: getProcedure
      tags: [procedures]
      summary: Get procedure
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Procedure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
        "404":
          description: Not found
    patch:
      operationId: updateProcedure
      tags: [procedures]
      summary: Update procedure
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcedureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
        "404":
          description: Not found
    delete:
      operationId: deleteProcedure
      tags: [procedures]
      summary: Delete procedure
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── SCHEDULED CALLS ───────────────────────────────────────────
...
  /call-templates:
    get:
      operationId: listCallTemplates
      tags: [templates]
      summary: List call templates
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallTemplate"
    post:
      operationId: createCallTemplate
      tags: [templates]
      summary: Create call template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTemplate"
  /call-templates/{id}:
    get:
      operationId: getCallTemplate
      tags: [templates]
      summary: Get call template
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTemplate"
        "404":
          description: Not found
    patch:
      operationId: updateCallTemplate
      tags: [templates]
      summary: Update call template
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallTemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTemplate"
        "404":
          description: Not found
    delete:
      operationId: deleteCallTemplate
      tags: [templates]
      summary: Delete call template
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────────
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List staff users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Provision a staff user account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  # ── VAPI WEBHOOK ──────────────────────────────────────────────
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, dateOfBirth, phone, createdAt]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
        phone: { type: string }
        email: { type: ["string", "null"] }
        mrn: { type: ["string", "null"] }
        language: { type: string, default: "en" }
        notes: { type: ["string", "null"] }
        ehrPatientId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Procedure:
      type: object
      required: [id, patientId, procedureName, scheduledDate, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        procedureName: { type: string }
        procedureCode: { type: ["string", "null"] }
        scheduledDate: { type: string, format: date-time }
        facility: { type: ["string", "null"] }
        surgeon: { type: ["string", "null"] }
        arrivalTime: { type: ["string", "null"] }
        specialInstructions: { type: ["string", "null"] }
        ehrProcedureId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        patient:
          $ref: "#/components/schemas/Patient"
    ProcedureInput:
      type: object
      required: [patientId, procedureName, scheduledDate]
      properties:
        patientId: { type: integer }
        procedureName: { type: string }
        procedureCode: { type: string }
        scheduledDate: { type: string, format: date-time }
        facility: { type: string }
        surgeon: { type: string }
        arrivalTime: { type: string }
        specialInstructions: { type: string }
        ehrProcedureId: { type: string }
    ProcedureUpdate:
      type: object
      properties:
        procedureName: { type: string }
        procedureCode: { type: string }
        scheduledDate: { type: string }
        facility: { type: string }
        surgeon: { type: string }
        arrivalTime: { type: string }
        specialInstructions: { type: string }

    # ── SCHEDULED CALL ──────────────────────────────────────────
...
    CallTemplate:
      type: object
      required: [id, callType, name, systemPrompt, language, active, createdAt]
      properties:
        id: { type: integer }
        callType:
          type: string
          enum: [pre_op_history, pre_op_instructions, post_op_24h, post_op_72h, post_op_2wk]
        name: { type: string }
        systemPrompt: { type: string }
        questions: { type: ["string", "null"] }
        language: { type: string, default: "en" }
        active: { type: boolean }
        vapiAssistantId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CallTemplateInput:
      type: object
      required: [callType, name, systemPrompt, language]
      properties:
        callType: { type: string, enum: [pre_op_history, pre_op_instructions, post_op_24h, post_op_72h, post_op_2wk] }
        name: { type: string }
        systemPrompt: { type: string }
        questions: { type: string }
        language: { type: string }
        active: { type: boolean }
        vapiAssistantId: { type: string }
    CallTemplateUpdate:
      type: object
      properties:
        name: { type: string }
        systemPrompt: { type: string }
        questions: { type: string }
        language: { type: string }
        active: { type: boolean }
        vapiAssistantId: { type: string }

    # ── DASHBOARD ───────────────────────────────────────────────
...
    User:
      type: object
      required: [id, clerkUserId, email, role, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        email: { type: string }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        role:
          type: string
          enum: [admin, coordinator]
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string, enum: [admin, coordinator] }
    UserUpdate:
      type: object
      properties:
        role: { type: string, enum: [admin, coordinator] }
        active: { type: boolean }

    # ── VAPI WEBHOOK ────────────────────────────────────────────
...